UFAL/Created job to generate preview for every item or for a specific one (#887)
* Created job to generate preview for every item or for a specific one.

* Added docs, fixed checkstyle, prettying the code

* Added integration tests for File Preview job
