+++ dspace-api/src/main/java/org/dspace/content/service/PreviewContentService.java
@@ -7,6 +7,8 @@
  */
 package org.dspace.content.service;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
@@ -14,8 +16,10 @@ import java.util.UUID;
 
 import org.dspace.authorize.AuthorizeException;
 import org.dspace.content.Bitstream;
+import org.dspace.content.Item;
 import org.dspace.content.PreviewContent;
 import org.dspace.core.Context;
+import org.dspace.util.FileInfo;
 
 /**
  * Service interface class for the PreviewContent object.
@@ -83,10 +87,10 @@ public interface PreviewContentService {
      * Find all preview content based on bitstream that are the root directory.
      *
      * @param context        DSpace context
-     * @param bitstream_id   The ID of the bitstream
+     * @param bitstream      The bitstream to get bitstream UUID
      * @throws SQLException  If a database error occurs
      */
-    List<PreviewContent> findRootByBitstream(Context context, UUID bitstream_id) throws SQLException;
+    List<PreviewContent> hasPreview(Context context, Bitstream bitstream) throws SQLException;
 
     /**
      * Find all preview contents from database.
@@ -95,4 +99,59 @@ public interface PreviewContentService {
      * @throws SQLException  If a database error occurs
      */
     List<PreviewContent> findAll(Context context) throws SQLException;
+
+    /**
+     * Find out if the bitstream could be previewed
+     *
+     * @param context DSpace context object
+     * @param bitstream check if this bitstream could be previewed
+     * @return true if the bitstream could be previewed, false otherwise
+     */
+    boolean canPreview(Context context, Bitstream bitstream) throws SQLException, AuthorizeException;
+
+    /**
+     * Return converted ZIP file content into FileInfo classes.
+     *
+     * @param context DSpace context object
+     * @param bitstream ZIP file bitstream
+     * @return List of FileInfo classes where is wrapped ZIP file content
+     */
+    List<FileInfo> getFilePreviewContent(Context context, Bitstream bitstream)
+            throws SQLException, AuthorizeException, IOException;
+
+    /**
+     * Create preview content from file info for bitstream.
+     *
+     * @param context   DSpace context object
+     * @param bitstream bitstream
+     * @param fi        file info
+     * @return          created preview content
+     * @throws SQLException If database error is occurred
+     */
+    PreviewContent createPreviewContent(Context context, Bitstream bitstream, FileInfo fi) throws SQLException;
+
+    /**
+     * Compose download URL for calling `MetadataBitstreamController` to download single file or
+     * all files as a single ZIP file.
+     */
+    String composePreviewURL(Context context, Item item, Bitstream bitstream, String contextPath) throws SQLException;
+
+    /**
+     * Create file info from preview content.
+     *
+     * @param pc  preview content
+     * @return    created file info
+     */
+    FileInfo createFileInfo(PreviewContent pc);
+
+    /**
+     * Convert InputStream of the ZIP file into FileInfo classes.
+     *
+     * @param context DSpace context object
+     * @param bitstream previewing bitstream
+     * @param inputStream content of the zip file
+     * @return List of FileInfo classes where is wrapped ZIP file content
+     */
+    List<FileInfo> processInputStreamToFilePreview(Context context, Bitstream bitstream, InputStream inputStream)
+            throws SQLException, IOException;
 }
