+++ dspace-api/src/test/data/dspaceFolder/config/spring/api/scripts.xml
@@ -101,4 +101,9 @@
         <property name="dspaceRunnableClass" value="org.dspace.app.healthreport.HealthReport"/>
     </bean>
 
+    <bean id="file-preview" class="org.dspace.scripts.filepreview.FilePreviewConfiguration" primary="true">
+        <property name="description" value="Generate the file preview"/>
+        <property name="dspaceRunnableClass" value="org.dspace.scripts.filepreview.FilePreview"/>
+    </bean>
+
 </beans>
