+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/PreviewContentServiceImplIT.java
@@ -125,7 +125,7 @@ public class PreviewContentServiceImplIT extends AbstractControllerIntegrationTe
     @Test
     public void testFindRootByBitstream() throws Exception {
         List<PreviewContent> previewContentList =
-                previewContentService.findRootByBitstream(context, bitstream1.getID());
+                previewContentService.hasPreview(context, bitstream1);
         Assert.assertEquals(previewContentList.size(), 1);
         Assert.assertEquals(previewContent1.getID(), previewContentList.get(0).getID());
     }
